feat(create_community_post): create_community_post working
create_community_post done, and testing is working. Next up, will refactor all the code, to instantiate a driver when a youtube class is instantiated, logging into Youtube by either getting in an email & password, cookies, or chrome profile path. Then, pass that chromium instance to each function, and the default function starting point is at the home page on Youtube (https://youtube.com), so that user doesn't have to pass credentials on each call. After that, create a .kill method, to kill the chromium driver.